+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,9 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **All components**
+  * Increased `compileSdkVersion` to 29 (Android Q)
+
 # 4.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v3.0.0...v4.0.0)